chore: move verification functions into a separate service
