+++ backend/src/email/services/email.service.ts
@@ -1,11 +1,9 @@
 import { Transaction } from 'sequelize'
 import logger from '@core/logger'
-import dns from 'dns'
-import AWS from 'aws-sdk'
 import { CSVParams } from '@core/types'
 
 import { ChannelType } from '@core/constants'
-import { Campaign, ProtectedMessage, User } from '@core/models'
+import { Campaign, ProtectedMessage } from '@core/models'
 import {
   MailService,
   CampaignService,
@@ -16,9 +14,7 @@ import {
 import { MailToSend, CampaignDetails } from '@core/interfaces'
 
 import { EmailTemplate, EmailMessage, VerifiedEmail } from '@email/models'
-import { EmailTemplateService } from '@email/services'
-
-const ses = new AWS.SES({ region: 'eu-central-1' })
+import { EmailTemplateService, CustomDomainService } from '@email/services'
 
 /**
  * Gets a message's parameters
@@ -284,61 +280,6 @@ const uploadProtectedCompleteOnChunk = ({
 }
 
 /**
- * Verifies if the cname records are in the email's domain dns
- * @throws Error if verification fails
- */
-// SWTODO: move to a separate service. Not sure about the service name yet
-const verifyCnames = async (
-  tokens: Array<string>,
-  email: string
-): Promise<void> => {
-  // get email domain
-  const domain = email.slice(email.lastIndexOf('@') + 1)
-  const cnames = tokens.map((token) => `${token}._domainkey.${domain}`)
-  try {
-    for (const cname of cnames) {
-      await dns.promises.resolve(cname, 'CNAME')
-    }
-  } catch (e) {
-    throw new Error(`Verification of dkim records failed for ${email}`)
-  }
-}
-
-/**
- * Verifies if the cname records are in the email's domain dns
- * @throws Error if the domain's dns do not have the cname records
- */
-const verifyEmailWithAWS = async (email: string): Promise<Array<string>> => {
-  // Get the dkim attributes for the email address
-  const params = {
-    Identities: [email],
-  }
-  const { DkimAttributes } = await ses
-    .getIdentityDkimAttributes(params)
-    .promise()
-  if (DkimAttributes[email]?.DkimVerificationStatus! !== 'Success') {
-    throw new Error(`Email address not verified by AWS SES. email=${email}`)
-  }
-  return DkimAttributes[email]?.DkimTokens!
-}
-
-/**
- * Verifies if the email provided matches the user's
- * @throws Error if the emails doesn't match
- */
-const verifyEmail = async (email: string, userId: number): Promise<void> => {
-  // Verify email address that is provided
-  const user = await User.findOne({ where: { id: userId } })
-  if (user === null) throw new Error(`Failed to find user id: ${userId}`)
-
-  if (user.email !== email) {
-    throw new Error(
-      `From email address not allowed. User's email: ${user.email}, given: ${email} `
-    )
-  }
-}
-
-/**
  * Verifies the from email address in different ways:
  * 1. Against the user's email address
  * 2. With AWS to ensure that we can use the email address to send
@@ -349,9 +290,9 @@ const verifyFromEmailAddress = async (
   email: string,
   userId: number
 ): Promise<void> => {
-  await verifyEmail(email, userId)
-  const DkimTokens = await verifyEmailWithAWS(email)
-  await verifyCnames(DkimTokens, email)
+  await CustomDomainService.verifyEmail(email, userId)
+  const DkimTokens = await CustomDomainService.verifyEmailWithAWS(email)
+  await CustomDomainService.verifyCnames(DkimTokens, email)
 }
 
 /**
