+++ backend/src/email/services/index.ts
@@ -1,3 +1,4 @@
 export * from './email.service'
 export * from './email-template.service'
 export * from './email-stats.service'
+export * from './custom-domain.service'
